Fixed pose_server and tracks imu package

+++ mission_planner_stack/pose_server/include/pose_server/Estimator.h
@@ -79,7 +79,7 @@ namespace kraken_core
       /*
        * Update time
        */
-      const int _time;
+      const float _time;
       
   };
